LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-strict_dynamic_javascript_uri-expected.txt

CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains an invalid source: ''strict-dynamic''. It will be ignored.
CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
Script injected via `javascript:` URIs are not allowed with `strict-dynamic`.


Harness Error (TIMEOUT), message = null

TIMEOUT Script injected via `javascript:` URIs are not allowed with `strict-dynamic`. Test timed out
